+++ packages/mesh-hydra/src/utils/index.ts
@@ -1,1 +1,2 @@
+export * from "./http";
 export * from "./parse-http-error";
